requirements.txt
asttokens==2.4.1
boltons==25.0.0
click==8.1.8
cloudpickle==3.1.1
cmake==3.31.4
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coverage==7.6.10
cvxopt==1.3.2
cycler==0.12.1
Cython==3.0.12
debtcollector==3.0.0
debugpy==1.8.9
decorator==5.1.1
dictdiffer==0.9.0
diff-pdf-visually==1.8.0
dit==1.5
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fonttools==4.55.0
igraph==0.11.8
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jpype1==1.5.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lattices==0.3.5
leidenalg==0.10.2
llvmlite==0.44.0
mantel==2.2.2
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
narwhals==1.25.1
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pillow==11.0.0
platformdirs==4.3.6
plotly==6.0.0
PLTable==1.1.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pybind11==2.13.6
pycddlib==3.0.2
PyGenStability==0.2.3
Pygments==2.18.0
pyparsing==3.2.0
pypoman==1.2.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-louvain==0.16
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shap==0.47.2
six==1.16.0
slicer==0.0.8
stack-data==0.6.3
statsmodels==0.14.4
texttable==1.7.0
threadpoolctl==3.5.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
wrapt==1.17.2
